refactor: move Client testing to test_client.py
